reset styles and added emotion
