+++ package.json
@@ -3,13 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/css": "^11.1.3",
     "@fluent/bundle": "^0.17.0",
     "@fluent/langneg": "^0.6.0",
     "@fluent/react": "^0.14.0",
-    "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.0.1",
-    "@mui/styled-engine": "^5.0.0",
-    "@mui/styled-engine-sc": "^5.0.0",
     "@reduxjs/toolkit": "^1.6.1",
     "@types/jest": "^27.0.2",
     "@types/node": "^16.10.1",
@@ -17,6 +14,7 @@
     "@types/react-dom": "^17.0.9",
     "@types/socket.io-client": "^3.0.0",
     "@types/styled-components": "^5.1.14",
+    "emotion-theming": "^11.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.5",
@@ -24,7 +22,6 @@
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "socket.io-client": "^4.2.0",
-    "styled-components": "^5.3.1",
     "typescript": "^4.4.3",
     "web-vitals": "^1.0.1",
     "yarn": "^1.22.11"
