--- src/App.css
@@ -1,11 +1,0 @@
-html,
-body,
-#root {
-  height: 100%;
-}
-
-.App {
-  display: flex;
-  flex-grow: 1;
-  height: 100vh;
-}
