+++ src/App.tsx
@@ -1,65 +1,14 @@
-import './App.css';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { Home, VideoChat } from './containers';
+import { Login, VideoChat } from './containers';
 import { useSelector } from 'react-redux';
-import { RootState } from './store/store';
-import { createTheme, ThemeProvider } from '@mui/material';
-
-const SafeRoute = (props: any) => {
-  const username = useSelector(
-    (state: RootState) => state.application.username,
-  );
-  if (!username) return <Route {...props} component={Home} />;
-  return <Route {...props} />;
-};
+import { usernameSelector } from './store/selectors';
 
 function App() {
-  const customTheme = createTheme({
-    palette: {
-      mode: 'dark',
-      primary: {
-        light: '#afa',
-        main: '#99e5df',
-        dark: '#afa',
-        contrastText: '#fff',
-      },
-      secondary: {
-        light: '#afa',
-        main: '#f11',
-        dark: '#faf',
-        contrastText: '#faf',
-      },
-    },
-    components: {
-      // Name of the component
-      MuiButton: {
-        styleOverrides: {
-          // Name of the slot
-          root: {
-            // Some CSS
-            fontSize: '1rem',
-            backgroundColor: '#357f78',
-            ':hover': {
-              backgroundColor: '#1a746b',
-            },
-            color: 'white',
-            outlineColor: 'white',
-          },
-        },
-      },
-    },
-  });
+  const username = useSelector(usernameSelector);
+  if (!username) {
+    return <Login />;
+  }
 
-  return (
-    <ThemeProvider theme={customTheme}>
-      <BrowserRouter>
-        <Switch>
-          <Route exact path="/" component={Home} />
-          <SafeRoute path="/video" component={VideoChat} />
-        </Switch>
-      </BrowserRouter>
-    </ThemeProvider>
-  );
+  return <VideoChat />;
 }
 
 export default App;
