+++ src/containers/index.ts
@@ -1,2 +1,2 @@
-export { default as Home } from './Home'
+export { default as Login } from './Login'
 export { default as VideoChat } from './VideoChat'
