+++ src/store/selectors.tsx
@@ -1,0 +1,4 @@
+import { RootState } from './store';
+
+export const usernameSelector = (state: RootState) =>
+  state.application.username;
